Moved internals inside a pointer reference.
What
===
Moved internal fields of `Static` into an internal type `static` that's
stored on a pointer on `Static`.

Why
===
So that those internal types are always referenced when the outer type
is copied. This means the main type can be copied and used as a value
type everywhere and the internal fields will never be copied.

Notes
===
The internal fields right now are maps which technically wouldn't matter
if they were copied, but I wanted to use this pattern.