Fix bug where custom EventHandlers were ignored.
What
===
Changed the `RenderEventHandler` method so that it renders using the
`Renderer` and passes through the `EventHandler` that it receives,
rather than passing the `Renderer` the default `logEvent`
`EventHandler`.

Why
===
It was a bug that the `EventHandler` argument was not used. This
function exists to allow the caller to specify the `EventHandler` given
to the `Renderer`, but it wasn't being passed through and the `logEvent`
`EventHandler`.