Entire refactor for simplification.
What
===
Entire refactor for simplification. This library now only handles
building, and can take any Go web app that uses a http.Handler. You
provide the handler and the list of paths (routes) to build, and it goes
and does that.

Why
===
The library was too much a framework and was trying to orchestrate too
much. It required you to write an app a certain way, where as this new
design encourages you to write a web app any way you want, and then this
will help you turn that into a static website.