Refactored responseBuffer into responseWriter.
What
===
Refactored `responseBuffer` into `responseWriter`, removing the tight
coupling to an underlying `bytes.Buffer` and instead relying on the
`io.Writer` interface to pass through the bytes that need writing.

Why
===
It was unnecessary for the `http.ResponseWriter` implementation to be
tied to any implementation of a buffer since the `responseBuffer` was
just storing the data in memory then doing a `WriteTo` a file which is
also an `io.Writer`. This way the page data is not being held in memory
unnecessarily and is piped directly to the file output. This also made
testing much simpler.