Fixing example tests for consistency.
What
===
Fixing example tests for consistency by reducing them to a single path
that gets built.

Why
===
The `Build` example was failing most of the time because the output
order could not be guaranteed due to concurrency. The same changes were
made to the `BuildSingle` so it continues to similar to the `Build`
example and is easier to compare.