Added recording of the HTTP Status Code.
What
===
Added recording of the HTTP Status Code.

Why
===
In the future the status code will be included in events.

And
===
In the process I cleaned up the tests for the responseWriter. I would
have done this in a separate commit but I got too deep with the two
things intertwined.