Added HTTP status code passed out for visibility.
What
===
Added HTTP status code passed out for visibility.

Why
===
An error may not occur but it may be useful for the user of the package
to know what the status code was that was returned by the http.Handler
during building.

Example: If the handler is a default http.ServeMux and the handler does
not have a route for the path, no error will occur but the status code
will be 404. This may be completely valid and the package doesn't make
any assumptions that a framework might, but this change gives the caller
the option to do something with it.