Make target to update the README coverage badge.
What
===
Added a make target to update the README coverage badge by running go
test with coverage, extracting the percentage and then replacing that in
the `README.md`.

Why
===
So that coverage is visible in the README. I did it using a simple Perl
because I didn't want to setup dependencies on coveralls.