Remove Chmod from tests.
What
===
Remove Chmod from tests where it was being used to simulate the package
from being unable to create a file at a path that already existed and
for which we did not have permission to write, and replaced it with
logic to create a directory with the same name as the file to be written.

Why
===
The package is designed for use on Linux, OS X and Windows, and Chmod is
not supported on Windows.

Signed-off-by: Leigh McCulloch <leigh@mcchouse.com>