Swapped nil checks for no-op defaults.
What
===
Swapped nil checks on EventHandler for a no-op default event handler
that does nothing.

Why
===
The Go Best Practices 2016 article linked below, recommended applying
sensible defaults over testing code for nils. By doing so code that uses
the event handler does not need to be concerned with whether it is
available or not, if one is not needed it can be an empty function.

https://peter.bourgon.org/go-best-practices-2016/#program-design

This may allow for compiler optimization since the compiler may know that
a variable is always nil and can remove out those calls. I don't know
that for sure but I'd be interested to find out.

Signed-off-by: Leigh McCulloch <leigh@mcchouse.com>