Don't build examples on Windows.
What
===
Don't build examples on Windows.

Why
===
The examples include file paths in their output and since path
separators are different on Windows, building and running the examples
on Windows would cause CI builds to fail.

There's no good solution to this, but since the examples are only used
for documentation and there are already tests we don't lose any testing
rigor by disabling them on the Windows builds.