Converted tests into blackbox tests. (#1)
What
===
Replaced `package static` with `package static_test` in tests that are
testing entirely exported functionality.

Why
===
Blackbox testing is preferred as it results in tests that use the
package in the exact same way that another developer would use the
package.

Tests were originally written this way but it was changed when it was
discovered that the go cover tool was not including blackbox tests in
the coverage percentage.

According to golang/go#15795 the tests were not being included in the
package as because the package had been included into the blackbox tests
using the dot syntax (`import "."`). Importing using the fully qualified
package name fixes this, which is how the blackbox testing imports has
been implemented in this commit.